requirements.txt
Bottleneck==1.3.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.1
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.7.3
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.25
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.6.0
